chore: clean up from pr review
