--- schema/api/v0/store/store.proto
@@ -1,28 +1,0 @@
-syntax = "proto3";
-
-package vorpal.store.v0;
-
-service StoreService {
-    rpc Exists (StoreRequest) returns (StoreExistsResponse);
-    rpc Pull (StoreRequest) returns (stream StorePullResponse);
-}
-
-enum StoreKind {
-    UNKNOWN_STORE_KIND = 0;
-    ARTIFACT = 1;
-    ARTIFACT_SOURCE = 2;
-}
-
-message StoreRequest {
-    StoreKind kind = 1;
-    string hash = 2;
-    string name = 3;
-}
-
-message StoreExistsResponse {
-    bool exists = 1;
-}
-
-message StorePullResponse {
-    bytes data = 1;
-}
